rgw/sfs: dbconn: convert Storage to shared_ptr
This ensures we only have one storage object, and everything that needs to
use it does so via StorageRef, which is a std::shared_ptr.  Storage has been
renamed StorageImpl, to make sure I caught all the previous uses of Storage
elsewhere.  If we want to change this to a different type in future, it
should be possible to do so just by changing the type of StorageRef in
dbconn.h and dbconn.cc - so long as the new type behaves like a pointer,
everything Should Just Work(TM).

Signed-off-by: Tim Serong <tserong@suse.com>