[experimental] rgw/sfs: dbconn: add storage_pool
This gives one Storage object per thread, created on demand as a copy of
the first Storage object created in the DBConn constructor.  I'm not sure
whether we need the lock in get_storage(), but suspect we probably do given
we're potentially modifying the storage_pool map.

Right now this is just a std::map<pthread_t, StorageImpl>, but I reckon I
might be able to change the latter to a struct or a pair or something that
_also_ included a pointer to each StorageImpl's raw sqlite3 connection,
via some trickery in the storage->on_open callback.  This would potentially
allow calling sqlite3_db_status() on each connection to get runtime stats.

We might also be able to get rid of the "sfs_db_lock" mutex I added earlier
if we're back to using multiple connections, becasue multiple connections
should avoid the "transaction inside transaction" problem I had earlier.

Signed-off-by: Tim Serong <tserong@suse.com>