cmake: use `nproc` if _num_cores is zero
The previous workaround set _num_cores to 1 if cmake thought
NUMBER_OF_LOGICAL_CORES was zero.  That worked fine, but it'd
be much better to take advantage of all possible cores.  This
version runs `nproc` which seems to work just fine under
qemu-aarch64 and gives us the actual number of CPUs.

Fixes: b6669263cacd56ff589529784066b9f637f79e87
Signed-off-by: Tim Serong <tserong@suse.com>